refactor tests, also fixes #139 (even though this issue was already closed)
