+++ fmriprep/workflows/tests/test_confounds.py
@@ -1,7 +1,7 @@
 ''' Testing module for fmriprep.workflows.confounds '''
 import logging
-from fmriprep.workflows.confounds import init_bold_confs_wf
-from .utilities import TestWorkflow
+from ..confounds import init_bold_confs_wf
+from ...utils.testing import TestWorkflow
 
 logging.disable(logging.INFO)  # don't print unnecessary msgs
